finalize deploy scripts
