+++ params.js
@@ -128,18 +128,19 @@ const MAIN_RIBBONOMICS_DIR = {
     WBTC: "0x8913EAb16a302dE3E498BbA39940e7A55c0B9325",
   },
   FUSE: {
-    COMPTROLLER: "",
-    INTEREST_RATE_MODEL: "",
-    UNDERLYING_NAME: "",
-    UNDERLYING_SYMBOL: "",
-    RESERVE_FACTOR_MANTISSA: "",
-    ADMIN_FEE_MANTISSA: "",
+    COMPTROLLER: "0xc1ee062d67e36abc20d17cf150d81834c129be44",
+    INTEREST_RATE_MODEL: "0x4ef29407a8dbca2f37b7107eab54d6f2a3f2ad60",
+    UNDERLYING_NAME:
+      "Ribbon's Birthday Cake Ribbon.fi rETH-THETA Gauge Deposit",
+    UNDERLYING_SYMBOL: "frETH-THETA-gauge-199",
+    RESERVE_FACTOR_MANTISSA: "100000000000000000",
+    ADMIN_FEE_MANTISSA: "0",
     CUSTOM_ORACLE: "0xe6BceFC66D6CbC4c555D53B817E738577477031E",
   },
   FUSE_REWARDS_DISTRIBUTOR: {
     START_TIME: 1651842000, // Fri May 06 2022 13:00:00 GMT+0000
-    RETH_THETA_CTOKEN: "",
-    USDC_CTOKEN: "",
+    RETH_THETA_CTOKEN: "0x69ac6d614fe2f4a6c6274231ba16155af5834b0d",
+    USDC_CTOKEN: "0x7f403ef3673c5ee110f970dfd020d781e4689a56",
     RETH_THETA_SUPPLIER_PCT: 10000, //100%
     USDC_SUPPLIER_PCT: 10000, //100%
   },
